Move data dashboard to separate route and controller
